requirements.txt
# Django and related packages
Django>=4.2.0
django-crispy-forms>=2.0
django-environ>=0.10.0

# Machine Learning and Data Processing
tensorflow>=2.12.0
numpy>=1.23.5
Pillow>=9.5.0
scikit-image>=0.20.0
lime>=0.2.0.1
matplotlib>=3.7.1
scipy>=1.10.1

# Web Interface and Frontend
django-tailwind>=3.6.0
fontawesome-free>=5.15.4

# Development and Security
python-dotenv>=1.0.0
gunicorn>=20.1.0
whitenoise>=6.4.0

# Image Processing
opencv-python>=4.7.0

# Database
django-mysql>=4.9.0
mysqlclient>=2.1.1
# psycopg2-binary>=2.9.6

# Cache and Performance
django-redis>=5.2.0
redis>=4.5.4

# Testing and Development
pytest>=7.3.1
pytest-django>=4.5.2
coverage>=7.2.3 